refactor: move card components into 'ui/cards' folder
