+++ src/ui/cards/ComplexCard/ComplexCard.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import { ComplexCard as ComplexCardType } from '@/types/Complex'
-import ComplexListCard from '@/components/Cards/Complex/ComplexListCard'
-import { ComplexTileCard } from '@/components/Cards/Complex/ComplexTileCard'
+import ComplexListCard from '@/ui/cards/ComplexCard/ComplexListCard'
+import { ComplexTileCard } from '@/ui/cards/ComplexCard/ComplexTileCard'
 
 interface Props {
   item: ComplexCardType
